fix(engine): add safety checks to savefile loading/storing (#1330)
* fix(engine): add safety check to savefile loading

If for some reason we try to access savefile for a player that should
have one, and it's either absent from the filesystem or has invalid
data, we now reject the login attempt under a generic "Please try again"
message.

Previously, there were no sanity checks on reading the savefile data, or
if we didn't find a savefile for a player which should have one
(`logout_time != null`). In theory, if something went wrong here, it
would be sent to the player as a completely reset tutorial island
account. Then, when they logged out, the new savefile would be written.

* fix(engine): prevent reset player saves

A last-ditch effort to prevent overwriting an existing save with one
that has been reset back to tutorial island.

* fix(engine): more robust tracking of wind-back

---------

Co-authored-by: locked-chest <179689881+sentientflare@users.noreply.github.com>